requirements.txt
zyte-api
requests
python-dotenv
